implemented common functions in simpleToolsUtils.js to avoid duplicates
