+++ renderer/components/input/subsetCreationTool.jsx
@@ -13,6 +13,7 @@ import { FilterMatchMode, FilterOperator } from "primereact/api"
 import { MultiSelect } from "primereact/multiselect"
 import { Utils as danfoUtils } from "danfojs-node"
 import SaveDataset from "../generalPurpose/saveDataset"
+import { cleanDataset, getColumnDataType, getData, getColumnOptions, getParentIDfolderPath, handleSelectedDatasetChange, updateListOfDatasets, updateTheColumnsTypes } from "./simpleToolsUtils"
 
 const dfd = require("danfojs-node")
 const dfdUtils = new danfoUtils()
@@ -38,45 +39,11 @@ const SubsetCreationTool = () => {
   const filterDisplay = "menu"
 
   /**
-   * To handle the change in the selected dataset, and update the columns options
-   * @param {Object} e - The event object
-   * @returns {Void}
-   */
-  const handleSelectedDatasetChange = async (e) => {
-    setSelectedDataset(globalData[e.target.value])
-  }
-
-  /**
-   * To update the list of datasets
-   * @returns {Void}
-   */
-  const updateListOfDatasets = () => {
-    let newDatasetList = []
-    Object.keys(globalData).forEach((key) => {
-      if (globalData[key].extension === "csv") {
-        newDatasetList.push({ name: globalData[key].name, object: globalData[key], key: key })
-      }
-    })
-    setListOfDatasets(newDatasetList)
-  }
-
-  /**
    * Hook that is called when the global data object is updated to update the list of datasets
    */
   useEffect(() => {
-    updateListOfDatasets()
+    updateListOfDatasets(globalData, selectedDataset, setListOfDatasets, setSelectedDataset)
   }, [globalData])
-
-  /**
-   * To get the data
-   * @returns {Promise} - The promise of the data
-   */
-  const getData = () => {
-    return new Promise((resolve) => {
-      let data = selectedDataset.loadDataFromDisk()
-      resolve(data)
-    })
-  }
 
   /**
    * This function initializes the filters
@@ -100,42 +67,11 @@ const SubsetCreationTool = () => {
   }
 
   /**
-   * Clean the dataset
-   * @param {DanfoJS.DataFrame} data - The data
-   * @returns {DanfoJS.DataFrame} - The cleaned dataset
-   */
-  const cleanDataset = (data) => {
-    if (data === null || data === undefined) {
-      return null
-    }
-
-    let newData = {}
-    let cleanedColumnNames = []
-    data.columns.forEach((column) => {
-      cleanedColumnNames.push(cleanString(column))
-    })
-
-    data.getColumnData.forEach((column, index) => {
-      let newColumn = []
-      column.forEach((value) => {
-        if (typeof value === "string") {
-          value = value.replace(/\s+/g, " ")
-          value = value.replace(/^[ '"]+|[ '"]+$|( ){2,}/g, "$1")
-        }
-        newColumn.push(value)
-      })
-      newData[cleanedColumnNames[index]] = newColumn
-    })
-    let df = new dfd.DataFrame(newData)
-    return df
-  }
-
-  /**
    * Hook that is called when the selected dataset is updated to update the columns infos
    */
   useEffect(() => {
     if (selectedDataset) {
-      getData().then((data) => {
+      getData(selectedDataset).then((data) => {
         data = cleanDataset(data)
 
         let columns = data.columns
@@ -166,39 +102,17 @@ const SubsetCreationTool = () => {
         MedDataObject.saveDatasetToDisk({ data: filteredData, filePath: selectedDataset.path, extension: selectedDataset.extension })
         setSelectedDataset(null)
       } else {
-        MedDataObject.saveDatasetToDisk({ data: filteredData, filePath: getParentIDfolderPath(selectedDataset) + newDatasetName + "." + newDatasetExtension, extension: newDatasetExtension })
+        MedDataObject.saveDatasetToDisk({
+          data: filteredData,
+          filePath: getParentIDfolderPath(selectedDataset, globalData) + newDatasetName + "." + newDatasetExtension,
+          extension: newDatasetExtension
+        })
       }
       MedDataObject.updateWorkspaceDataObject()
     } else {
       // As create/overwrite button are disabled while filtered data is null, the only error to throw here is when filteredData.length == dataset.length
       toast.error("No filter applied")
     }
-  }
-
-  /**
-   * This function is used to clean a string
-   * @param {string} string - The string to clean
-   * @returns the cleaned string
-   * @summary This function is used to clean a string from spaces and quotes
-   */
-  const cleanString = (string) => {
-    if (string.includes(" ") || string.includes('"')) {
-      string = string.replaceAll(" ", "")
-      string = string.replaceAll('"', "")
-    }
-    return string
-  }
-
-  /**
-   * To get the parent ID folder path
-   * @param {Object} dataset - The dataset
-   * @returns {String} - The parent ID folder path with a trailing separator
-   */
-  const getParentIDfolderPath = (dataset) => {
-    let parentID = dataset.parentID
-    let parentPath = globalData[parentID].path
-    let separator = MedDataObject.getPathSeparator()
-    return parentPath + separator
   }
 
   const clearFilter = () => {
@@ -239,17 +153,7 @@ const SubsetCreationTool = () => {
    * This hook is used to update the column types
    */
   useEffect(() => {
-    if (df !== null && df !== undefined) {
-      let newColumnTypes = {}
-      df.ctypes.$data.forEach((type, index) => {
-        if (df.nUnique(0).$data[index] < 10) {
-          // If the number of unique values is less than 10, then it is a category
-          type = "category"
-        }
-        newColumnTypes[df.columns[index]] = type
-      })
-      setColumnTypes(newColumnTypes)
-    }
+    updateTheColumnsTypes(df, setColumnTypes)
   }, [df])
 
   /**
@@ -344,37 +248,6 @@ const SubsetCreationTool = () => {
     return string[0].toUpperCase() + string.slice(1)
   }
 
-  /**
-   * This function is used to get the column data type
-   * @param {string} column - The column
-   * @returns {string} - The column data type
-   */
-  function getColumnDataType(column) {
-    if (columnTypes[column] === "int32" || columnTypes[column] === "float32") {
-      return "numeric"
-    } else if (columnTypes[column] === "string") {
-      return "text"
-    } else if (columnTypes[column] === "category") {
-      return "category"
-    } else if (columnTypes[column] === "bool") {
-      return "boolean"
-    }
-  }
-
-  /**
-   * This function is used to get the column options according to the column type
-   * @param {string} column - The column
-   * @returns {Object} - The column options
-   */
-  const getColumnOptions = (column) => {
-    let optionsToReturn = { showFilterMatchModes: true, showFilterMenu: true }
-    if (columnTypes[column] === "category") {
-      optionsToReturn.showFilterMatchModes = false
-      optionsToReturn.showFilterMenu = true
-    }
-    return optionsToReturn
-  }
-
   return (
     <>
       <Row className="simple-cleaning-set">
@@ -387,7 +260,7 @@ const SubsetCreationTool = () => {
             optionValue="key"
             className="w-100"
             value={selectedDataset ? selectedDataset.getUUID() : null}
-            onChange={handleSelectedDatasetChange}
+            onChange={(e) => handleSelectedDatasetChange(e, setSelectedDataset, globalData)}
           ></Dropdown>
 
           <Row style={{ display: "flex", justifyContent: "space-evenly", flexDirection: "row", marginTop: "0.5rem" }}>
@@ -412,8 +285,8 @@ const SubsetCreationTool = () => {
                 selectedDatasetColumns.map((column, index) => (
                   <Column
                     key={column}
-                    {...getColumnOptions(column)}
-                    dataType={getColumnDataType(column)}
+                    {...getColumnOptions(column, columnTypes)}
+                    dataType={getColumnDataType(column, columnTypes)}
                     field={String(column)}
                     sortable
                     filterPlaceholder={`Search by ${column}`}
